feat: proteccion de endpoints con roles
se agregaron validaciones de ownership mediante metodo estatico y nuevo claim en el token JWT, esto con la finalidad de proteger la informacion de los usuarios y no repetir endpoints en funciones que el staff podria resolver tambien

+++ routes/carrera.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 use App\Lectores\Controllers\CarreraController;
@@ -13,22 +14,22 @@ $carreraRepository = new CarreraRepository();
 $carreraService = new CarreraService($carreraRepository);
 $carreraController = new CarreraController($carreraService);
 
-$router->get("/carreras", function() use($carreraController) {
+$router->get("/carreras", withRole(['admin', 'auxiliar', 'lector'], function () use ($carreraController) {
     $carreraController->getAll();
-});
+}));
 
-$router->get("/carreras/{id}", function($id) use($carreraController) {
+$router->get("/carreras/{id}", withRole(['admin', 'auxiliar', 'lector'], function ($id) use ($carreraController) {
     $carreraController->getById($id);
-});
+}));
 
-$router->post("/carreras", function() use($carreraController) {
+$router->post("/carreras", withRole(['admin'], function () use ($carreraController) {
     $carreraController->createCarrera();
-});
+}));
 
-$router->patch("/carreras/{id}", function($id) use($carreraController) {
+$router->patch("/carreras/{id}", withRole(['admin'], function ($id) use ($carreraController) {
     $carreraController->updateCarrera($id);
-});
+}));
 
-$router->delete("/carreras/{id}", function($id) use($carreraController) {
+$router->delete("/carreras/{id}", withRole(['admin'], function ($id) use ($carreraController) {
     $carreraController->deleteCarrera($id);
-});
+}));
